+++ packages/amplify-graphql-transformer-core/src/utils/index.ts
@@ -1,5 +1,6 @@
 export { DirectiveWrapper } from './directive-wrapper';
 export { collectDirectives, collectDirectivesByTypeNames } from './type-map-utils';
 export { stripDirectives } from './strip-directives';
+export { getTable, getKeySchema } from './schema-utils';
 export { DEFAULT_SCHEMA_DEFINITION } from './defaultSchema';
 export { IAM_AUTH_ROLE_PARAMETER, IAM_UNAUTH_ROLE_PARAMETER } from './authType';