+++ packages/amplify-e2e-tests/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.7.11](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-tests@3.7.10...amplify-e2e-tests@3.7.11) (2022-02-18)
+
+
+
+## 7.6.21 (2022-02-16)
+
+**Note:** Version bump only for package amplify-e2e-tests
+
+
+
+
+
 ## [3.7.10](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-tests@3.7.9...amplify-e2e-tests@3.7.10) (2022-02-15)
 
 **Note:** Version bump only for package amplify-e2e-tests